Feature/562 url filter (#565)
* Adds vanity URL filtering to the URL manager

Implements filtering for the vanity URL column in the URL manager.
This allows users to quickly find specific vanity URLs by typing
a search term.

The filtering is case-insensitive and searches within the RowKey
property of the ShortUrlEntity.  The data grid is also rebound to
only show the filter list.

* Adds title filtering to URL manager

Implements filtering functionality for the "Title" column in the URL manager.

This allows users to search URLs based on their titles, enhancing the ability to quickly find specific URLs.

Also refactors the vanity filter clear handling.